added drawer
implemented the drawer component from react-navigator. works
simultaniously with the Stack navigator and gets nested in that. custom
entries to the drawer can be done in the customDrawer.js with the
DrawerItem components. added a proof-of-concept link to the THM
homepage. any other elements such as images, texts or anything else can
also be added there in order to style the drawer freely.