Navigate through screens logic
Implemented logic to decide which task will be displayed next. Takes care of user navigation so it displays the next or the previous task depending on the pressed button.

Also takes user's overall as well as the topic's progress into account so it won't navigate to out of bounds indexes.

Currently it loops, so when the user clicked through all tasks it starts again from the first one. This needs to be handled otherwise (display summary screen in the future).