Functionality and animations
Introduced new functionality to split up provided text into sentences. Developer can additionally provide a MAX_TEXT_LENGTH to further split sentences if they are too long. Otherwise the text is split at "." or "!" so we get complete sentences on the screen.

Basic animation implemented and changes to the OSA screen are made accordingly so the back and forward buttons are not always shown.