implemented logic for proceed buttons in taskscreens
the next button now only gets displayed in quiz tasks, when the task is
checked. the task check button is still more or less a dummy button.
later this button should execute the logic for checking the users input.
after that button is clicked the next button will appear. implemented a
barebones screen for the interactive tasks which also receives the
showNextButton callback. because the interactive tasks aren't
implemented yet, the screen just calls the callback immediately in the
constructor.