implemented logic for different styles of quizes
the quizscreens now reads an additional property of the quiztask:
"style". this style determines if the quit is going to be a multiple or
a single choice style quiz. wether it is or not it renders out radio or
checkbox buttons accordingly.
additionally the taskscreens now display the current topic at the top of
the task.