Content, animation and code fixes
Made some adjustments to the text, images and animations. Improved timing of animations and visibility of elements. Image cards now fade into a thm-green background and are visible throughout the current task. Animations can still be improved. Added some padding, shadow and boldness to the text so it should be easier to read.

Changed some of the pictures and included new ones. There are still unused ones in the folder.

Made some small changes to the text for a better user experience.