changes to top bar in osa screen
removed the home button in the osa screen since you can use androids
back button to get back to the main screen. removed next and previous buttons
next to the progressbar. placed the previous button where the home
button used to be and deleted the next button since it is redundant due
to most recent changes.