Added ExamplesTask and ExamplesScreen
Added a new kind of task and it's corresponding view to present media informatic projects from former / current students. The view is basically a grid that contains thumbnails. For now, there is a simple log output when an image is clicked, furhter logic has to be implemented. Also some cosmetical fixes and fixed a bug that caused abnormal behaviour when navigating to a previous task.