changes to video pause behaviour
video does not fully reset on pause anymore, but rather pauses at the
current timestamp and will continue at that point again when pressed on
play. that change made it possible to enable the seek slider while
paused, to go to a certain point in the videos while the video is in
paused state.