changed navigator
removed MainNavigator and letting everything be handled by the
DrawerNavigator now. This acts as the main navigator now and thus every
screen now has access to the navigation.openDrawer() method. A side
effect of this, is that if you go back from e.g. the osa screen to the
homescreen, and go back to the osa screen, the osa will continue at the
point of exiting. can be a positive side effect, if the osa is finished
though, you'll be prompted with a blank screen. this should be more
controlled. did not find a method to reset everything, since the
TASK_MANAGER.resetTaskManager() method did not work as desired.